requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-localflavor==3.0.1
django-tempus-dominus==5.1.2.13
djangorestframework==3.12.2
python-stdnum==1.14
pytz==2020.4
soupsieve==2.1
sqlparse==0.4.1
